Cleanup our source-code organization.
Like many of the other projects I examined we should move
all the compiler-specific functionality into its own file
for clarity.

Also renamed the main `eval.go` to match our package-name
which is a standard golang approach.

This updates #90.
